+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './pair';
 export * from './swap';
 export * from './farm';
+export * from './api';
